[counters] Add string support
To register a string counter, you need to use the API like with any other type : you call mono_counters_register (NAME, SECTION | MONO_COUNTER_STRING | ..., &value), counter must be of type char**. The only restriction being that when replacing the value of the string, you **HAVE TO** FIRST set the new value, THEN release the old one. Otherwise a race condition could occur, leading to sampling the counters accessing a freed memory region. To replace the value, you simply do value = "new string" or value = g_strdup_printf (...).

+++ mono/utils/mono-counters.h
@@ -67,7 +67,7 @@ typedef mono_bool (*CountersEnumCallback) (MonoCounter *counter, void *user_data
 
 MONO_API void mono_counters_foreach (CountersEnumCallback cb, void *user_data);
 
-MONO_API int mono_counters_sample (MonoCounter *counter, void *buffer, int buffer_size);
+MONO_API int mono_counters_sample (MonoCounter *counter, void *buffer, size_t buffer_size);
 
 MONO_API const char* mono_counter_get_name (MonoCounter *name);
 MONO_API int mono_counter_get_type (MonoCounter *counter);
